refactor: split side nav components
